requirements.txt
langchain
langchain-community
pandas
requests
faiss-cpu
numpy
transformers
sentence-transformers
python-dotenv
